Swapped networking plugin with renet.

+++ src/core/atmospherics/events.rs
@@ -1,16 +1,16 @@
 use crate::core::networking::resources::ReliableServerMessage;
 
 pub struct NetMapDisplayAtmospherics {
-    pub handle: u32,
+    pub handle: u64,
     pub message: ReliableServerMessage,
 }
 
 pub struct NetMapHoverAtmospherics {
-    pub handle: u32,
+    pub handle: u64,
     pub message: ReliableServerMessage,
 }
 
 pub struct NetAtmosphericsNotices {
-    pub handle: u32,
+    pub handle: u64,
     pub message: ReliableServerMessage,
 }
