+++ src/core/connected_player/resources.rs
@@ -4,6 +4,6 @@ use bevy_ecs::entity::Entity;
 
 #[derive(Default)]
 pub struct HandleToEntity {
-    pub map: HashMap<u32, Entity>,
-    pub inv_map: HashMap<Entity, u32>,
+    pub map: HashMap<u64, Entity>,
+    pub inv_map: HashMap<Entity, u64>,
 }
