+++ src/core/console_commands/events.rs
@@ -3,12 +3,12 @@ use bevy_ecs::entity::Entity;
 use crate::core::networking::resources::{ConsoleCommandVariantValues, ReliableServerMessage};
 
 pub struct NetConsoleCommands {
-    pub handle: u32,
+    pub handle: u64,
     pub message: ReliableServerMessage,
 }
 
 pub struct InputConsoleCommand {
-    pub handle_option: Option<u32>,
+    pub handle_option: Option<u64>,
     pub entity: Entity,
     pub command_name: String,
     pub command_arguments: Vec<ConsoleCommandVariantValues>,
