+++ src/core/entity/functions/unload_entity_for_player.rs
@@ -3,7 +3,7 @@ use bevy_ecs::{entity::Entity, event::EventWriter};
 use crate::core::{entity::events::NetUnloadEntity, networking::resources::ReliableServerMessage};
 
 pub fn unload_entity(
-    player_handle: u32,
+    player_handle: u64,
     entity_id: Entity,
     net_unload_entity: &mut EventWriter<NetUnloadEntity>,
     unload_entirely: bool,
