+++ src/core/health/events.rs
@@ -30,6 +30,6 @@ pub struct Attack {
 }
 
 pub struct NetHealthUpdate {
-    pub handle: u32,
+    pub handle: u64,
     pub message: ReliableServerMessage,
 }
